+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [7.0.1] - 2025-06-24
+
+### Added
+
+- Glass Shield
+- Infectious Enthusiasm
+- Inside Ropes
+- Musical Accompaniment
+- Protect Companion
+- Ray of Frost
+- Read Aura
+- Read the Air
+- Rousing Splash
+- Take Root
+- Tangle Vine
+
+### Fixed
+
+- Infinite Effects w/ Dice So Nice!
+
 ## [7.0.0] - 2025-06-17
 
 Foundry Virtual Tabletop - Version 13 Support!
@@ -345,7 +365,8 @@ I don't remember all the changes that I made before I stopped working on this fo
 - Auto Self-Applied Effects (This is disabled if PF2e Toolbelt is doing this as well)
 - Swashbuckler Panache
 
-[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.0.0...HEAD
+[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.0.1...HEAD
+[7.0.1]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.0.0...v7.0.1
 [7.0.0]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.6.2...v7.0.0
 [6.6.2]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.6.1...v6.6.2
 [6.6.1]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.6.0...v6.6.1